Edit Week 2 and 3
I suggest to create only heroes-list component and display the list of heroes in HTML right away. I think it wouldn't be a 1 week's worth of work just to generate 2 components and add nothing more (since showing the list with *ngFor is part of week 3 at the moment). In week 3 they could add Angular Material to project. In previous years we started with Angular Material too lately. Students already had some styles applied in their projects and they had to transform them to Angular Material styles. It would be better to start with it right at the beginning so they can use this library and get familiar with it as they develop their apps in time.