Edit Week 4, 5, 6 and 7
Week 4: If the FE and BE applications should work together students need to implement BE calls on FE too. Content of the week 15 can't be a BONUS excercise as it is core angular feature. BONUS could be to get in touch with rxjs operators and understand how they work together with signals.
Week 5: To learn navigation it would be better to add new component as part of this week so they can play a bit more with navigation.
Week 6: Once we have navigation to details view done we can add functionality to edit and remove hero details. BE should be prepared already at that time. Two-way data binding can't be part of the week 7 because we need to use it in week 6 already.
Week 7: We can learn Angular signals instead of two-way data binding here.